requirements.txt
fastapi==0.63.0
priority==1.3.0
pydantic==1.8.1
python-multipart==0.0.5
python-decouple==3.4
